feat: add some oop design patterns
